finished some test

+++ exercises/error_handling/errors4.rs
@@ -14,7 +14,7 @@ enum CreationError {
 
 impl PositiveNonzeroInteger {
     fn new(value: i64) -> Result<PositiveNonzeroInteger, CreationError> {
-        Ok(PositiveNonzeroInteger(value as u64))
+        Ok(PositiveNonzeroInteger(value))
     }
 }
